+++ apiserver/client/client.go
@@ -948,6 +948,8 @@ func (c *Client) ShareEnvironment(args params.ModifyEnvironUsers) (result params
 }
 
 // GetAnnotations returns annotations about a given entity.
+// This API is now deprecated - "Annotations" client should be used instead.
+// TODO(anastasiamac) remove for Juju 2.x
 func (c *Client) GetAnnotations(args params.GetAnnotations) (params.GetAnnotationsResults, error) {
 	nothing := params.GetAnnotationsResults{}
 	tag, err := c.parseEntityTag(args.Tag)
@@ -958,7 +960,7 @@ func (c *Client) GetAnnotations(args params.GetAnnotations) (params.GetAnnotatio
 	if err != nil {
 		return nothing, errors.Trace(err)
 	}
-	ann, err := state.Annotations(entity, c.api.state)
+	ann, err := c.api.state.Annotations(entity)
 	if err != nil {
 		return nothing, errors.Trace(err)
 	}
@@ -989,6 +991,8 @@ func (c *Client) findEntity(tag names.Tag) (state.GlobalEntity, error) {
 }
 
 // SetAnnotations stores annotations about a given entity.
+// This API is now deprecated - "Annotations" client should be used instead.
+// TODO(anastasiamac) remove for Juju 2.x
 func (c *Client) SetAnnotations(args params.SetAnnotations) error {
 	tag, err := c.parseEntityTag(args.Tag)
 	if err != nil {
@@ -998,7 +1002,7 @@ func (c *Client) SetAnnotations(args params.SetAnnotations) error {
 	if err != nil {
 		return errors.Trace(err)
 	}
-	return state.SetAnnotations(entity, c.api.state, args.Pairs)
+	return c.api.state.SetAnnotations(entity, args.Pairs)
 }
 
 // parseSettingsCompatible parses setting strings in a way that is
